+++ src/Ritten/Workflows/DotNet/DotNetJob.cs
@@ -1,5 +1,4 @@
 using Ritten.CodeCoverage;
-using Ritten.DotNet;
 using Ritten.Engine;
 using Ritten.Engine.Workflows;
 using Ritten.Reporting;